# docs/swagger.yaml
basePath: /
definitions:
...
  dto.PhotoArchiveResponse:
    properties:
      category:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      image_path:
        type: string
      taken_date:
        type: string
      title:
        type: string
    type: object
  dto.TheatreCreateRequest:
    properties:
      notes:
        type: string
      play_title:
        maxLength: 255
        minLength: 1
        type: string
      premiere_date:
        type: string
      theatre_name:
        maxLength: 255
        minLength: 1
        type: string
    required:
    - play_title
    - theatre_name
    type: object
  dto.TheatreResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      notes:
        type: string
      play_title:
        type: string
      premiere_date:
        type: string
      theatre_name:
        type: string
    type: object
...
  handler.JSONResponse:
    properties:
      data: {}
      message:
        type: string
      status:
        type: string
    type: object
host: localhost:8081
info:
  contact: {}
  description: API for Hemra Şirow literary and biography website
  title: Literary Backend API
  version: "1.0"
paths:
...
  /api/photo-archive:
    get:
      description: List all photo archive entries
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.PhotoArchiveResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.JSONResponse'
      summary: List photo archive
    post:
      consumes:
      - application/json
      - multipart/form-data
      description: Create a new photo archive entry. Supports JSON or multipart/form-data
        with optional image upload.
      parameters:
      - description: Title
        in: formData
        name: title
        required: true
        type: string
      - description: Description
        in: formData
        name: description
        type: string
      - description: Category
        in: formData
        name: category
        type: string
      - description: Image file
        in: formData
        name: image
        type: file
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.PhotoArchiveResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.JSONResponse'
      summary: Create photo archive entry
  /api/photo-archive/{id}:
    delete:
      description: Delete an existing photo archive entry by id
      parameters:
      - description: Photo archive ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.JSONResponse'
      summary: Delete photo archive entry
    get:
      description: Get a single photo archive entry by id
      parameters:
      - description: Photo archive ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PhotoArchiveResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.JSONResponse'
      summary: Get photo archive
    put:
      consumes:
      - application/json
      - multipart/form-data
      description: Update an existing photo archive entry. Supports JSON or multipart/form-data
        with optional image upload.
      parameters:
      - description: Photo archive ID
        in: path
        name: id
        required: true
        type: string
      - description: Title
        in: formData
        name: title
        type: string
      - description: Description
        in: formData
        name: description
        type: string
      - description: Image file
        in: formData
        name: image
        type: file
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PhotoArchiveResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.JSONResponse'
      summary: Update photo archive entry
  /api/theatre-productions:
    get:
      description: List all theatre productions
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.TheatreResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.JSONResponse'
      summary: List theatre productions
    post:
      consumes:
      - application/json
      description: Create a new theatre production
      parameters:
      - description: Theatre production data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.TheatreCreateRequest'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.TheatreResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.JSONResponse'
      summary: Create theatre production
  /api/theatre-productions/{id}:
    delete:
      description: Delete an existing theatre production by id
      parameters:
      - description: Theatre production ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.JSONResponse'
      summary: Delete theatre production
    get:
      description: Get a single theatre production by id
      parameters:
      - description: Theatre production ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TheatreResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.JSONResponse'
      summary: Get theatre production
    put:
      consumes:
      - application/json
      description: Update an existing theatre production
      parameters:
      - description: Theatre production ID
        in: path
        name: id
        required: true
        type: string
      - description: Theatre production data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.TheatreCreateRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TheatreResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.JSONResponse'
      summary: Update theatre production
...
schemes:
- http
- https
swagger: "2.0"